handler/requirements.txt
minio==7.1.12
elastic-transport==8.4.0
elasticsearch==8.4.3
kafka-python==2.0.2
path==16.5.0
fastapi==0.85.0
supervisor==4.2.4